LayoutTests/platform/ios-simulator/accessibility/tables-lists-expected.txt

item 1
list item 1
table item 1	table item 2	table item 3
landmark 1
This test makes sure that items in lists, tables and landmarks have different traits on them, because they're contained in different items.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS listItem.traits != initialItem.traits is true
PASS listItem.traits != tableItem.traits is true
PASS listItem.traits != landmarkItem.traits is true
PASS tableItem.traits != initialItem.traits is true
PASS tableItem.traits != landmarkItem.traits is true
PASS landmarkItem.traits != initialItem.traits is true
PASS successfullyParsed is true

TEST COMPLETE
